added mute functionality

+++ main.py
@@ -48,18 +48,21 @@ def msg(db_msg_id, mongo_db=db):
 
 
 # прикрепить тэг к пользователю
-@bot.message_handler(commands=['add_tags'])
+@bot.message_handler(commands=['tag_add'])
 def handle_tag_adding_start(tg_message):
     try:
         check_user_for_admin_right(tg_message.from_user, tg_message.chat)
 
         response_msg = bot.send_message(tg_message.chat.id, "Укажи студента и тэг")
-        bot.register_next_step_handler(response_msg, handle_tag_adding_end)
+        bot.register_next_step_handler(response_msg, handle_tag_adding_end, tg_message.from_user)
     except NoAdminRights as e:
         bot.send_message(tg_message.chat.id, e.msg)
 
 
-def handle_tag_adding_end(tg_message):
+def handle_tag_adding_end(tg_message, user):
+    if tg_message.from_user.id != user.id:
+        bot.register_next_step_handler(tg_message, handle_tag_adding_end, user)
+        return
     try:
         username, tags = tg_message.text.split(' ', 1)
 
@@ -78,17 +81,20 @@ def handle_tag_adding_end(tg_message):
         bot.send_message(tg_message.chat.id, e.msg)
 
 
-@bot.message_handler(commands=['remove_tags'])
+@bot.message_handler(commands=['tag_remove'])
 def remove_tags_start(tg_message):
     try:
         check_user_for_admin_right(tg_message.from_user, tg_message.chat)
-        response_msg = bot.send_message(tg_message.chat.id, "Укажите пользователя и тэги")
-        bot.register_next_step_handler(response_msg, remove_tags_end)
+        response_msg = bot.send_message(tg_message.chat.id, "Укажите пользователя и тэги через пробел")
+        bot.register_next_step_handler(response_msg, remove_tags_end, tg_message.from_user)
     except NoAdminRights as e:
         bot.send_message(tg_message.chat.id, e.msg)
 
 
-def remove_tags_end(tg_message):
+def remove_tags_end(tg_message, user):
+    if tg_message.from_user.id != user.id:
+        bot.register_next_step_handler(tg_message, remove_tags_end, user)
+        return
     try:
         username, tags = tg_message.text.split(' ', 1)
 
@@ -107,6 +113,66 @@ def remove_tags_end(tg_message):
         bot.send_message(tg_message.chat.id, e.msg)
 
 
+# замутить пользователей по тэгам
+@bot.message_handler(commands=['tag_mute'])
+def handle_tag_mute_start(tg_message):
+    if tg_message.chat.type != "supergroup":
+        bot.send_message(tg_message.chat.id, "Ошибка: данная команда возможна только в супергруппе")
+        return
+    try:
+        check_user_for_admin_right(tg_message.from_user, tg_message.chat)
+        bot.send_message(tg_message.chat.id, "Укажи тэги через пробел")
+        bot.register_next_step_handler(tg_message, handle_tag_mute_end, tg_message.from_user)
+    except NoAdminRights as e:
+        bot.send_message(tg_message.chat.id, e.msg)
+
+
+def handle_tag_mute_end(tg_message, user):
+    if tg_message.from_user.id != user.id:
+        bot.register_next_step_handler(tg_message, handle_tag_mute_end, user)
+        return
+    try:
+        tags = tg_message.text.split(" ")
+
+        users = user_tags.get_users_with_tags(tags, "", tg_message.chat, bot, db)
+        for user in users:
+            bot.restrict_chat_member(tg_message.chat.id, user.id, False)
+
+        bot.send_message(tg_message.chat.id, "Успех")
+    except user_tags.UserTagsException as e:
+        bot.send_message(tg_message.chat.id, e.msg)
+
+
+# замутить пользователей по тэгам
+@bot.message_handler(commands=['tag_unmute'])
+def handle_tag_unmute_start(tg_message):
+    if tg_message.chat.type != "supergroup":
+        bot.send_message(tg_message.chat.id, "Ошибка: данная команда возможна только в супергруппе")
+        return
+    try:
+        check_user_for_admin_right(tg_message.from_user, tg_message.chat)
+        bot.send_message(tg_message.chat.id, "Укажи тэги через пробел")
+        bot.register_next_step_handler(tg_message, handle_tag_unmute_end, tg_message.from_user)
+    except NoAdminRights as e:
+        bot.send_message(tg_message.chat.id, e.msg)
+
+
+def handle_tag_unmute_end(tg_message, user):
+    if tg_message.from_user.id != user.id:
+        bot.register_next_step_handler(tg_message, handle_tag_unmute_end, user)
+        return
+    try:
+        tags = tg_message.text.split(" ")
+
+        users = user_tags.get_users_with_tags(tags, "", tg_message.chat, bot, db)
+        for user in users:
+            bot.restrict_chat_member(tg_message.chat.id, user.id, True, True, True, True, True, True)
+
+        bot.send_message(tg_message.chat.id, "Успех")
+    except user_tags.UserTagsException as e:
+        bot.send_message(tg_message.chat.id, e.msg)
+
+
 # уведомить пользователей с хотя бы одним из указанных тэгов
 @bot.message_handler(commands=['tag'])
 def handle_tag__start(tg_message):
@@ -114,7 +180,7 @@ def handle_tag__start(tg_message):
 
 
 # уведомить пользователей со всеми указанными тэгов
-@bot.message_handler(commands=['tag_all'])
+@bot.message_handler(commands=['tag_inclusive'])
 def handle_tag_all_start(tg_message):
     handle_tag_middle(tg_message, "all")
 
@@ -122,19 +188,22 @@ def handle_tag_all_start(tg_message):
 def handle_tag_middle(tg_message, _type):
     if tg_message.reply_to_message is None:
         bot.send_message(tg_message.chat.id,
-                         "Ответь этой командой на сообщение, под которым будут отмечены пользователи")
+                         "Ответь этой командой через пробел на сообщение, под которым будут отмечены пользователи")
         return
 
     # удалить сообщение с командой, оно нам больше не нужно
     bot.delete_message(tg_message.chat.id, tg_message.message_id)
 
     # отправить подсказку и перенаправить ответ
-    response_msg = bot.send_message(tg_message.chat.id, "Укажи тэги")
-    bot.register_next_step_handler(response_msg, handle_tag_end, tg_message.reply_to_message, response_msg, _type)
+    response_msg = bot.send_message(tg_message.chat.id, "Укажи тэги через пробел")
+    bot.register_next_step_handler(response_msg, handle_tag_end, tg_message.reply_to_message, response_msg, _type, tg_message.from_user)
 
 
 # главный обработчик уведомления пользователей по тэгам
-def handle_tag_end(tg_message, original_message, info_message, _type):
+def handle_tag_end(tg_message, original_message, info_message, _type, user):
+    if tg_message.from_user.id != user.id:
+        bot.register_next_step_handler(tg_message, handle_tag_end, user)
+        return
     try:
         # удалить сообщение "Укажи тэги"
         bot.delete_message(info_message.chat.id, info_message.message_id)
@@ -172,19 +241,22 @@ def handle_tag_end(tg_message, original_message, info_message, _type):
 
 
 # создать тэг в бд
-@bot.message_handler(commands=['create_tag'])
+@bot.message_handler(commands=['tag_create'])
 def handle_create_tag_start(tg_message):
     try:
         check_user_for_admin_right(tg_message.from_user, tg_message.chat)
 
-        response_msg = bot.send_message(tg_message.chat.id, "Укажи название и описание тэга")
-        bot.register_next_step_handler(response_msg, handle_create_tag_end)
+        response_msg = bot.send_message(tg_message.chat.id, "Укажи название и описание тэга через пробел")
+        bot.register_next_step_handler(response_msg, handle_create_tag_end, tg_message.from_user)
 
     except NoAdminRights as e:
         bot.send_message(tg_message.chat.id, e.msg)
 
 
-def handle_create_tag_end(tg_message):
+def handle_create_tag_end(tg_message, user):
+    if tg_message.from_user.id != user.id:
+        bot.register_next_step_handler(tg_message, handle_create_tag_end, user)
+        return
     try:
         index = tg_message.text.find(" ")
         if index != -1:
@@ -202,17 +274,20 @@ def handle_create_tag_end(tg_message):
 
 
 # удалить тэг из бд и открепить всех пользователей от него
-@bot.message_handler(commands=['destroy_tag'])
+@bot.message_handler(commands=['tag_destroy'])
 def handle_destroy_tag_start(tg_message):
     try:
         check_user_for_admin_right(tg_message.from_user, tg_message.chat)
         response_msg = bot.send_message(tg_message.chat.id, "Укажи имя тэга")
-        bot.register_next_step_handler(response_msg, handle_destroy_tag_end)
+        bot.register_next_step_handler(response_msg, handle_destroy_tag_end, tg_message.from_user)
     except NoAdminRights as e:
         bot.send_message(tg_message.chat.id, e.msg)
 
 
-def handle_destroy_tag_end(tg_message):
+def handle_destroy_tag_end(tg_message, user):
+    if tg_message.from_user.id != user.id:
+        bot.register_next_step_handler(tg_message, handle_destroy_tag_end, user)
+        return
     try:
         user_tags.destroy_tag(tg_message.text, db)
     except user_tags.UserTagsException as e:
@@ -220,13 +295,16 @@ def handle_destroy_tag_end(tg_message):
 
 
 # получить тэги прикрепленные к даному пользователю
-@bot.message_handler(commands=['user_tags'])
-def handle_user_tags_begin(tg_message):
-    response_msg = bot.send_message(tg_message.chat.id, "Укажите пользователя")
-    bot.register_next_step_handler(response_msg, handle_user_tags_end)
+@bot.message_handler(commands=['tag_user'])
+def handle_user_tags_start(tg_message):
+    response_msg = bot.send_message(tg_message.chat.id, "Укажите пользователя через отметку")
+    bot.register_next_step_handler(response_msg, handle_user_tags_end, tg_message.from_user)
 
 
-def handle_user_tags_end(tg_message):
+def handle_user_tags_end(tg_message, user):
+    if tg_message.from_user.id != user.id:
+        bot.register_next_step_handler(tg_message, handle_user_tags_end, user)
+        return
     try:
         username = tg_message.text[1:]
 
@@ -314,12 +392,15 @@ def send_message(tg_message):
         text = 'В ответе на это сообщение пришлите текст нового пункта'
         response_msg = bot.send_message(tg_message.chat.id, text, parse_mode="html")
 
-        def _add_to_faq(new_text):
+        def _add_to_faq(new_text, user):
+            if new_text.from_user.id != user.id:
+                bot.register_next_step_handler(new_text, _add_to_faq, tg_message.from_user)
+                return
             add_to_faq(db, new_text.text)
             text = 'FAQ обновлен'
             bot.send_message(tg_message.chat.id, text, parse_mode="html")
 
-        bot.register_next_step_handler(response_msg, _add_to_faq)
+        bot.register_next_step_handler(response_msg, _add_to_faq, tg_message.from_user)
     except NoAdminRights as e:
         bot.send_message(tg_message.chat.id, e.msg)
 
@@ -329,11 +410,13 @@ def send_message(tg_message):
 def send_message(tg_message):
     try:
         check_user_for_admin_right(tg_message.from_user, tg_message.chat)
-        text = 'В ответе на это сообщение пришлите номер удаляемого пункта'
+        text = 'Укажи номер удаляемого пункта'
         response_msg = bot.send_message(tg_message.chat.id, text, parse_mode="html")
 
-        def _remove_from_faq(new_text):
-
+        def _remove_from_faq(new_text, user):
+            if new_text.from_user.id != user.id:
+                bot.register_next_step_handler(new_text, _remove_from_faq, tg_message.from_user)
+                return
             try:
                 remove_from_faq(db, int(new_text.text))
                 text = 'FAQ обновлен'
@@ -341,7 +424,7 @@ def send_message(tg_message):
                 text = 'Ошибка: ожидалось число. FAQ оставлен без изменений.'
             bot.send_message(tg_message.chat.id, text, parse_mode="html")
 
-        bot.register_next_step_handler(response_msg, _remove_from_faq)
+        bot.register_next_step_handler(response_msg, _remove_from_faq, tg_message.from_user)
     except NoAdminRights as e:
         bot.send_message(tg_message.chat.id, e.msg)
 
@@ -360,23 +443,26 @@ def send_message(tg_message):
 
 # получить информацию о преподавателях
 @bot.message_handler(commands=['teacher'])
-def get_teacher_info_prepare(message):
+def get_teacher_info_prepare(tg_message):
     _msg = bot.send_message(
-        message.chat.id, 'Отправьте имя преподавателя.', parse_mode="html"
+        tg_message.chat.id, 'Отправьте имя преподавателя.', parse_mode="html"
     )
-    bot.register_next_step_handler(_msg, get_teacher_info)
+    bot.register_next_step_handler(_msg, get_teacher_info, tg_message.from_user)
 
 
-def get_teacher_info(message):
-    if message.text.lower() in ['bruh']:
+def get_teacher_info(tg_message, user):
+    if tg_message.from_user.id != user.id:
+        bot.register_next_step_handler(tg_message, get_teacher_info, user)
+        return
+    if tg_message.text.lower() in ['bruh']:
         result = '''Страница 1 из 1
 
 <b>Имя:</b> <code>Bruhский Bruh Bruhович</code>
 <b>Телефон: </b><code>22505</code>
 <b>E-mail: </b><code>herzen_bruhs@hello.world</code>'''
     else:
-        result = teachers_parser.parse_teacher(message.text)
-    bot.send_message(message.chat.id, result, parse_mode="html")
+        result = teachers_parser.parse_teacher(tg_message.text)
+    bot.send_message(tg_message.chat.id, result, parse_mode="html")
 
 
 # handle new chat members
@@ -385,8 +471,11 @@ def handle_new_chat_members(tg_message):
     if tg_message.new_chat_members is not None:
         for newUser in tg_message.new_chat_members:
             try:
-                user_tags.register_user(newUser, db)
-                bot.send_message(tg_message.chat.id, "Добро пожаловать, " + newUser.first_name + "!")
+                if newUser.id == bot.get_me().id:
+                    bot.send_message(tg_message.chat.id, "Всем привет!")
+                else:
+                    user_tags.register_user(newUser, db)
+                    bot.send_message(tg_message.chat.id, "Добро пожаловать, " + newUser.first_name + "!")
             except user_tags.UserTagsException as e:
                 if e.msg == "Error: user already exists":
                     bot.send_message(tg_message.chat.id, "Добро пожаловать обратно, " + newUser.first_name + "!")
